+++ packages/plugin-admin-api/test/legacy/settings-api.spec.js
@@ -50,6 +50,7 @@ describe("settings api", () => {
           },
         },
         files: {
+          enabled: true,
           babelRegister: {
             enabled: false,
             options: {},